POAP drop - multichain monetization setup
